+++ app/lib/config/init.dart
@@ -183,10 +183,10 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart) async {
 
   if (appStart) {
     if (defaultTargetPlatform == TargetPlatform.macOS) {
-      final files = await getPendingFiles();
-      if (files.isNotEmpty) {
+      final pendingFiles = await getPendingFiles();
+      if (pendingFiles.isNotEmpty) {
         await ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
-          args: files,
+          args: pendingFiles,
         ));
       }
 
@@ -195,6 +195,23 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart) async {
         ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
           args: files,
         ));
+        ref.redux(homePageControllerProvider).dispatch(ChangeTabAction(HomeTab.send));
+      });
+
+      final pendingStrings = await getPendingStrings();
+      if (pendingStrings.isNotEmpty) {
+        for (final string in pendingStrings) {
+          ref.redux(selectedSendingFilesProvider).dispatch(AddMessageAction(message: string));
+          ref.redux(homePageControllerProvider).dispatch(ChangeTabAction(HomeTab.send));
+        }
+      }
+
+      // handle future dropped strings
+      getPendingStringsStream().listen((pendingStrings) {
+        for (final string in pendingStrings) {
+          ref.redux(selectedSendingFilesProvider).dispatch(AddMessageAction(message: string));
+          ref.redux(homePageControllerProvider).dispatch(ChangeTabAction(HomeTab.send));
+        }
       });
     } else {
       final args = ref.read(appArgumentsProvider);